Flytt funksjon for generering av personresultat til egen fil
