+++ src/test/integrasjonstester/kotlin/no/nav/familie/ba/sak/datagenerator/vilkårsvurdering/Vilkårsvurdering.kt
@@ -2,6 +2,7 @@ package no.nav.familie.ba.sak.datagenerator.vilkårsvurdering
 
 import io.mockk.mockk
 import no.nav.familie.ba.sak.datagenerator.lagPerson
+import no.nav.familie.ba.sak.datagenerator.lagPersonResultatAvOverstyrteResultater
 import no.nav.familie.ba.sak.datagenerator.lagVilkårResultat
 import no.nav.familie.ba.sak.kjerne.autovedtak.fødselshendelse.Resultat
 import no.nav.familie.ba.sak.kjerne.behandling.domene.Behandling
